+++ dotcom-rendering/src/web/components/CalloutEmbed/FormFields/FileUpload.tsx
@@ -2,9 +2,9 @@
 import { css } from '@emotion/react';
 import { space, text, textSans } from '@guardian/source-foundations';
 import { useState } from 'react';
-import type { CampaignFieldFile } from '../../../types/content';
-import { stringifyFileBase64 } from '../../lib/stringifyFileBase64';
-import { FieldLabel } from '../Callout/FormFields/FieldLabel';
+import type { CampaignFieldFile } from '../../../../types/content';
+import { stringifyFileBase64 } from '../../../lib/stringifyFileBase64';
+import { FieldLabel } from './FieldLabel';
 
 const fileUploadInputStyles = css`
 	padding-top: 10px;
